Implement loyalty tier evaluation draft

+++ AutoWash-Pro/backend/src/main/java/com/autowash/backend/reward/controller/RewardController.java
@@ -126,7 +126,7 @@ public class RewardController {
      */
     @GetMapping("/redeemable")
     public ResponseEntity<List<RewardResponseDTO>> getRedeemableRewards(
-            @RequestParam Long customerId,
+            @RequestParam Integer customerId,
             @RequestParam String vehicleType
     ) {
         List<RewardResponseDTO> rewards =
