+++ AutoWash-Pro/backend/src/main/java/com/autowash/backend/reward/dto/RedeemRewardRequestDTO.java
@@ -15,7 +15,7 @@ public record RedeemRewardRequestDTO(
         }
     */
     @NotNull(message = "Yeu cau ID khach hang")
-    Long customerId,
+    Integer customerId,
     /*
         @NotBlank:
              Không được null
